chore(TT-831): make es batch settings configurable
Make the batch size settings for writing to Elasticsearch configurable
in the environment. Also make the coalesce and repartition sizes
configurable, since those are related to how many workers write to
Elasticsearch.

Disable automatic post-write index refresh so that Elasticsearch does
not become unavailable after we write a bunch of data.