chore(models): exclude old diagnoses fields
Exclude deprecated diagnoses fields that have been removed from the
models, but are still present in the original source data.

If we do not exclude those fields, the mutation indexer's output will
not match the mappings and indexing will fail. We can remove the
exclusions once those fields have been cleaned out of the graph index.