feat(multiply-df): don't repartition case or maf
Don't repartition the intermediate case or MAF dataframes.
Feed them into the subsequent stages as-is.

Repartitioning the MAF dataframe takes hours with the data scaled up
and I am not convinced repartitioning by ssm_id is good for some of the
things we do with that dataframe later on. Skip repartitioning and
see what happens.