feat(multiply-df): make score centric ssm-only
Rename score_centric to ssm_score_centric and remove CNVs therefrom.
Only index SSMs like that for now.

If we need similar support for CNVs, we'll add a separate index in
the future. Avoid the challenges we've encountered with mixing SSM
and CNV data in the same index in the past.