feat(spark-conf): make more things configurable (#260)
* feat(spark-conf): add spark conf args

Add command line arguments for two Spark configuration options
relevant to building large volumes of data (shuffle partitions
and auto-broadcast join threshold).

Set the default auto-broadcast join threshold to work around
failures we sometimes observe when building the GISTIC data.

* feat(spark-conf): configure maxResultSize

Make spark.driver.maxResultSize configurable. Leave the default at 1g.

This shouldn't be necessary for the amount of data we're building now,
but it may become necessary as we index more data in the future.

* feat(spark-conf): make graph index configurable

Allow configuring the name of the graph index to read (--graph-index).
Keep the gdc_from_graph alias as the default.

* feat(spark-conf): shuffle less aggressively

Drop the default shuffle partitions down to 1024.

This is still higher than the Spark default (200), but not as high as
the huge value I'd set when doing scaling tests. I think this is a good
guess at the right value...?

* feat(spark-conf): tidy argument lookups

Refactor the lookups of argument values from the argparse Namespace.
Remove some duplication and save a few lines.