DEV-101 update index naming for dave-ca phase 1
Implement the new index name conventions for DAVE-CA Phase 1.

Add a `--study-label` parameter that determines whether the indices
should be represented as open- or controlled-access. Name the built
indices accordingly.

Make the `--build-type` and `--build-version` settings do nothing for
now. We may use them in the future to enable some automated builds
based on the relased-projects table.