DEV-1268: Add case field selector service (#347)
DEV-1268: Add a service for selecting all case fields for a given viz index.
This service finds all case related fields in an index's mapping from the normalizer. This will allow us to select only the fields we need to retrieve from the case index when building the case data associated with a viz index (or multiple if they have the same case data).