DEV-1342: Remove TestConfig from conftest & optimize spark session for tests. (#358)
This removes the TestConfig from the conftest file which moves the integration test suite closer to being able to remove its dependency on TestConfig/BaseConfig. The ConfigAdapter is maintained to allow for backwards compatibility but should be the ONLY way the BaseConfig is utilized as it will be removed when all refs to the BaseConfig are removed.

In doing the above, it was reveled that the test_base_builder_cast_boolean function was breaking fixture functionality by forcing pytest to reevaluate the various *centric_dfs session fixtures by calling request.getfixturevalue. This would cause the test to try and reload the es indices causing an error to be raised.

Also this speeds up the integration and unit tests by utilizing some advanced configurations of the spark session.