DEV-3084: implement cnv_change_5_category logic (#465)
* DEV-3084: implement cnv_change_5_category logic

A few unit tests for cnv_change had to be chnaged, as they led to ploidy
values < 1, which is not possible in the real world. Ploidy values < 1
leads to different number of non-null rows between cnv_change and
cnv_change_5_category.

A follow-up change will be made to change the cnv_id calculation, which is required in order to avoid overwriting of documents in certain cases.